gitflow-feature-stash: progress-bar
Foi adicionado uma barra de progresso para o candidato consiga saber quantas candidaturas aquele job já possui.

+++ src/main/java/br/edu/utfpr/servicebook/model/dto/JobRequestDetailsDTO.java
@@ -20,6 +20,6 @@ public class JobRequestDetailsDTO implements Serializable {
     private String description;
     private Long totalCandidates;
     private String textualDate;
-    private Long amountOfCandidates;
+    private Integer amountOfCandidates;
 
 }
